+++ src/index.js
@@ -1,40 +1,49 @@
 'use strict'
 
-var parseHTML = require('./parse-html')
-var KEY_PREFIX = '_set-dom-'
-var NODE_INDEX = KEY_PREFIX + 'index'
-var NODE_MOUNTED = KEY_PREFIX + 'mounted'
-var ELEMENT_TYPE = window.Node.ELEMENT_NODE
-var DOCUMENT_TYPE = window.Node.DOCUMENT_NODE
 setDOM.KEY = 'data-key'
 setDOM.IGNORE = 'data-ignore'
+setDOM.CHECKSUM = 'data-checksum'
+var parseHTML = require('./parse-html')
+var KEY_PREFIX = '_set-dom-'
+var NODE_MOUNTED = KEY_PREFIX + 'mounted'
+var ELEMENT_TYPE = 1
+var DOCUMENT_TYPE = 9
+var DOCUMENT_FRAGMENT_TYPE = 11
 
+// Expose api.
 module.exports = setDOM
 
 /**
  * @description
  * Updates existing dom to match a new dom.
  *
- * @param {Node} prev - The html entity to update.
- * @param {String|Node} next - The updated html(entity).
+ * @param {Node} oldNode - The html entity to update.
+ * @param {String|Node} newNode - The updated html(entity).
  */
-function setDOM (prev, next) {
+function setDOM (oldNode, newNode) {
   // Ensure a realish dom node is provided.
-  assert(prev && prev.nodeType, 'You must provide a valid node to update.')
+  assert(oldNode && oldNode.nodeType, 'You must provide a valid node to update.')
 
   // Alias document element with document.
-  if (prev.nodeType === DOCUMENT_TYPE) prev = prev.documentElement
+  if (oldNode.nodeType === DOCUMENT_TYPE) oldNode = oldNode.documentElement
 
-  // If a string was provided we will parse it as dom.
-  if (typeof next === 'string') next = parseHTML(next, prev.nodeName)
-
-  // Update the node.
-  setNode(prev, next)
+  // Document Fragments don't have attributes, so no need to look at checksums, ignored, attributes, or node replacement.
+  if (newNode.nodeType === DOCUMENT_FRAGMENT_TYPE) {
+    // Simply update all children (and subchildren).
+    setChildNodes(oldNode, newNode)
+  } else {
+    // Otherwise we diff the entire old node.
+    setNode(oldNode, typeof newNode === 'string'
+      // If a string was provided we will parse it as dom.
+      ? parseHTML(newNode, oldNode.nodeName)
+      : newNode
+    )
+  }
 
   // Trigger mount events on initial set.
-  if (!prev[NODE_MOUNTED]) {
-    prev[NODE_MOUNTED] = true
-    mount(prev)
+  if (!oldNode[NODE_MOUNTED]) {
+    oldNode[NODE_MOUNTED] = true
+    mount(oldNode)
   }
 }
 
@@ -43,43 +52,42 @@ function setDOM (prev, next) {
  * @description
  * Updates a specific htmlNode and does whatever it takes to convert it to another one.
  *
- * @param {Node} prev - The previous HTMLNode.
- * @param {Node} next - The updated HTMLNode.
+ * @param {Node} oldNode - The previous HTMLNode.
+ * @param {Node} newNode - The updated HTMLNode.
  */
-function setNode (prev, next) {
-  if (prev.nodeType === next.nodeType) {
+function setNode (oldNode, newNode) {
+  if (oldNode.nodeType === newNode.nodeType) {
     // Handle regular element node updates.
-    if (prev.nodeType === ELEMENT_TYPE) {
-      // Ignore elements that explicity choose not to be diffed.
-      if (!(prev.attributes[setDOM.IGNORE] && next.attributes[setDOM.IGNORE])) {
-        // Update all children (and subchildren).
-        setChildNodes(prev, prev.childNodes, next.childNodes)
+    if (oldNode.nodeType === ELEMENT_TYPE) {
+      // Checks if nodes are equal before diffing.
+      if (isEqualNode(oldNode, newNode)) return
 
-        // Update the elements attributes / tagName.
-        if (prev.nodeName === next.nodeName) {
-          // If we have the same nodename then we can directly update the attributes.
-          setAttributes(prev, prev.attributes, next.attributes)
-        } else {
-          // Otherwise clone the new node to use as the existing node.
-          var newPrev = next.cloneNode()
-          // Copy over all existing children from the original node.
-          while (prev.firstChild) newPrev.appendChild(prev.firstChild)
-          // Replace the original node with the new one with the right tag.
-          prev.parentNode.replaceChild(newPrev, prev)
-        }
+      // Update all children (and subchildren).
+      setChildNodes(oldNode, newNode)
+
+      // Update the elements attributes / tagName.
+      if (oldNode.nodeName === newNode.nodeName) {
+        // If we have the same nodename then we can directly update the attributes.
+        setAttributes(oldNode.attributes, newNode.attributes)
+      } else {
+        // Otherwise clone the new node to use as the existing node.
+        var newPrev = newNode.cloneNode()
+        // Copy over all existing children from the original node.
+        while (oldNode.firstChild) newPrev.appendChild(oldNode.firstChild)
+        // Replace the original node with the new one with the right tag.
+        oldNode.parentNode.replaceChild(newPrev, oldNode)
       }
     } else {
       // Handle other types of node updates (text/comments/etc).
       // If both are the same type of node we can update directly.
-      if (prev.nodeValue !== next.nodeValue) {
-        prev.nodeValue = next.nodeValue
+      if (oldNode.nodeValue !== newNode.nodeValue) {
+        oldNode.nodeValue = newNode.nodeValue
       }
     }
   } else {
     // we have to replace the node.
-    dismount(prev)
-    prev.parentNode.replaceChild(next, prev)
-    mount(next)
+    oldNode.parentNode.replaceChild(newNode, dismount(oldNode))
+    mount(newNode)
   }
 }
 
@@ -88,43 +96,34 @@ function setNode (prev, next) {
  * @description
  * Utility that will update one list of attributes to match another.
  *
- * @param {Node} parent - The current parentNode being updated.
- * @param {NamedNodeMap} prev - The previous attributes.
- * @param {NamedNodeMap} next - The updated attributes.
+ * @param {NamedNodeMap} oldAttributes - The previous attributes.
+ * @param {NamedNodeMap} newAttributes - The updated attributes.
  */
-function setAttributes (parent, prev, next) {
+function setAttributes (oldAttributes, newAttributes) {
   var i, a, b, ns, name
 
   // Remove old attributes.
-  for (i = prev.length; i--;) {
-    a = prev[i]
+  for (i = oldAttributes.length; i--;) {
+    a = oldAttributes[i]
     ns = a.namespaceURI
     name = a.localName
-    b = next.getNamedItemNS(ns, name)
-    if (!b) prev.removeNamedItemNS(ns, name)
+    b = newAttributes.getNamedItemNS(ns, name)
+    if (!b) oldAttributes.removeNamedItemNS(ns, name)
   }
 
   // Set new attributes.
-  for (i = next.length; i--;) {
-    a = next[i]
+  for (i = newAttributes.length; i--;) {
+    a = newAttributes[i]
     ns = a.namespaceURI
     name = a.localName
-    b = prev.getNamedItemNS(ns, name)
-    if (typeof parent[name] !== undefined) {
-      if(name === "checked" || name === "selected"){
-        parent[name] = "checked";
-      } else{
-        parent[name] = (name === "checked" && typeof a.value === "string") ? (a.value === "true" ? true : false): a.value
-      }
-    }
+    b = oldAttributes.getNamedItemNS(ns, name)
     if (!b) {
       // Add a new attribute.
-      next.removeNamedItemNS(ns, name)
-      prev.setNamedItemNS(a)
+      newAttributes.removeNamedItemNS(ns, name)
+      oldAttributes.setNamedItemNS(a)
     } else if (b.value !== a.value) {
       // Update existing attribute.
       b.value = a.value
-      parent.value = a.value
     }
   }
 }
@@ -132,79 +131,75 @@ function setAttributes (parent, prev, next) {
 /**
  * @private
  * @description
- * Utility that will update one list of childNodes to match another.
+ * Utility that will nodes childern to match another nodes children.
  *
- * @param {Node} parent - The current parentNode being updated.
- * @param {NodeList} prevChildNodes - The previous children.
- * @param {NodeList} nextChildNodes - The updated children.
+ * @param {Node} oldParent - The existing parent node.
+ * @param {Node} newParent - The new parent node.
  */
-function setChildNodes (parent, prevChildNodes, nextChildNodes) {
-  var key, a, b, newPosition, nextEl
+function setChildNodes (oldParent, newParent) {
+  var checkOld, oldKey, checkNew, newKey, foundNode, keyedNodes
+  var oldNode = oldParent.firstChild
+  var newNode = newParent.firstChild
+  var extra = 0
 
-  // Convert nodelists into a usuable map.
-  var prev = keyNodes(prevChildNodes)
-  var next = keyNodes(nextChildNodes)
+  // Extract keyed nodes from previous children and keep track of total count.
+  while (oldNode) {
+    extra++
+    checkOld = oldNode
+    oldKey = getKey(checkOld)
+    oldNode = oldNode.nextSibling
 
-  // Remove old nodes.
-  for (key in prev) {
-    if (next[key]) continue
-    // Trigger custom dismount event.
-    dismount(prev[key])
-    // Remove child from dom.
-    parent.removeChild(prev[key])
+    if (oldKey) {
+      if (!keyedNodes) keyedNodes = {}
+      keyedNodes[oldKey] = checkOld
+    }
   }
 
-  // Set new nodes.
-  for (key in next) {
-    a = prev[key]
-    b = next[key]
-    // Extract the position of the new node.
-    newPosition = b[NODE_INDEX]
+  // Loop over new nodes and perform updates.
+  oldNode = oldParent.firstChild
+  while (newNode) {
+    extra--
+    checkNew = newNode
+    newNode = newNode.nextSibling
 
-    if (a) {
-      // Update an existing node.
-      setNode(a, b)
-      // Check if the node has moved in the tree.
-      if (a[NODE_INDEX] === newPosition) continue
-      // Get the current element at the new position.
-      /* istanbul ignore next */
-      nextEl = prevChildNodes[newPosition] || null // TODO: figure out if || null is needed.
-      // Check if the node has already been properly positioned.
-      if (nextEl === a) continue
-      // Reposition node.
-      parent.insertBefore(a, nextEl)
+    if (keyedNodes && (newKey = getKey(checkNew)) && (foundNode = keyedNodes[newKey])) {
+      delete keyedNodes[newKey]
+      // If we have a key and it existed before we move the previous node to the new position if needed and diff it.
+      if (foundNode !== oldNode) {
+        oldParent.insertBefore(foundNode, oldNode)
+      } else {
+        oldNode = oldNode.nextSibling
+      }
+
+      setNode(foundNode, checkNew)
+    } else if (oldNode) {
+      checkOld = oldNode
+      oldNode = oldNode.nextSibling
+      if (getKey(checkOld)) {
+        // If the old child had a key we skip over it until the end.
+        oldParent.insertBefore(checkNew, checkOld)
+        mount(checkNew)
+      } else {
+        // Otherwise we diff the two non-keyed nodes.
+        setNode(checkOld, checkNew)
+      }
     } else {
-      // Get the current element at the new position.
-      nextEl = prevChildNodes[newPosition] || null
-      // Append the new node at the correct position.
-      parent.insertBefore(b, nextEl)
-      // Trigger custom mounted event.
-      mount(b)
+      // Finally if there was no old node we add the new node.
+      oldParent.appendChild(checkNew)
+      mount(checkNew)
     }
   }
-}
 
-/**
- * @private
- * @description
- * Converts a nodelist into a keyed map.
- * This is used for diffing while keeping elements with 'data-key' or 'id' if possible.
- *
- * @param {NodeList} childNodes - The childNodes to convert.
- * @return {Object}
- */
-function keyNodes (childNodes) {
-  var result = {}
-  var len = childNodes.length
-  var el
-
-  for (var i = 0; i < len; i++) {
-    el = childNodes[i]
-    el[NODE_INDEX] = i
-    result[getKey(el) || i] = el
+  // Remove old keyed nodes.
+  for (oldKey in keyedNodes) {
+    extra--
+    oldParent.removeChild(dismount(keyedNodes[oldKey]))
   }
 
-  return result
+  // If we have any remaining unkeyed nodes remove them from the end.
+  while (--extra >= 0) {
+    oldParent.removeChild(dismount(oldParent.lastChild))
+  }
 }
 
 /**
@@ -214,64 +209,101 @@ function keyNodes (childNodes) {
  * Uses 'data-key' if possible and falls back to 'id'.
  *
  * @param {Node} node - The node to get the key for.
- * @return {String}
+ * @return {string|void}
  */
 function getKey (node) {
   if (node.nodeType !== ELEMENT_TYPE) return
   var key = node.getAttribute(setDOM.KEY) || node.id
-  if (key) key = KEY_PREFIX + key
-  return key && KEY_PREFIX + key
+  if (key) return KEY_PREFIX + key
 }
 
 /**
- * Recursively trigger a mount event for a node and it's children.
+ * Checks if nodes are equal using the following by checking if
+ * they are both ignored, have the same checksum, or have the
+ * same contents.
  *
- * @param {Node} node - the initial node to be mounted.
+ * @param {Node} a - One of the nodes to compare.
+ * @param {Node} b - Another node to compare.
  */
-function mount (node) {
-  // Trigger mount event for this element if it has a key.
-  if (getKey(node)) dispatch(node, 'mount')
-
-  // Mount all children.
-  var child = node.firstChild
-  while (child) {
-    mount(child)
-    child = child.nextSibling
-  }
-}
-
-/**
- * Recursively trigger a dismount event for a node and it's children.
- *
- * @param {Node} node - the initial node to be dismounted.
- */
-function dismount (node) {
-  // Dismount all children.
-  var child = node.firstChild
-  while (child) {
-    dismount(child)
-    child = child.nextSibling
-  }
-
-  // Trigger dismount event for this element if it has a key.
-  if (getKey(node)) dispatch(node, 'dismount')
+function isEqualNode (a, b) {
+  return (
+    // Check if both nodes are ignored.
+    (isIgnored(a) && isIgnored(b)) ||
+    // Check if both nodes have the same checksum.
+    (getCheckSum(a) === getCheckSum(b)) ||
+    // Fall back to native isEqualNode check.
+    a.isEqualNode(b)
+  )
 }
 
 /**
  * @private
  * @description
- * Create and dispatch a custom event.
+ * Utility to try to pull a checksum attribute from an element.
+ * Uses 'data-checksum' or user specified checksum property.
  *
- * @param {Node} el - the node to dispatch the event for.
- * @param {String} type - the name of the event.
+ * @param {Node} node - The node to get the checksum for.
+ * @return {string|NaN}
  */
-function dispatch (el, type) {
-  var e = document.createEvent('Event')
-  var prop = { value: el }
-  e.initEvent(type, false, false)
-  Object.defineProperty(e, 'target', prop)
-  Object.defineProperty(e, 'srcElement', prop)
-  el.dispatchEvent(e)
+function getCheckSum (node) {
+  return node.getAttribute(setDOM.CHECKSUM) || NaN
+}
+
+/**
+ * @private
+ * @description
+ * Utility to try to check if an element should be ignored by the algorithm.
+ * Uses 'data-ignore' or user specified ignore property.
+ *
+ * @param {Node} node - The node to check if it should be ignored.
+ * @return {boolean}
+ */
+function isIgnored (node) {
+  return node.getAttribute(setDOM.IGNORE) != null
+}
+
+/**
+ * Dispatches a mount event for the given node and children.
+ *
+ * @param {Node} node - the node to mount.
+ * @return {node}
+ */
+function mount (node) {
+  return dispatch(node, 'mount')
+}
+
+/**
+ * Dispatches a dismount event for the given node and children.
+ *
+ * @param {Node} node - the node to dismount.
+ * @return {node}
+ */
+function dismount (node) {
+  return dispatch(node, 'dismount')
+}
+
+/**
+ * Recursively trigger an event for a node and it's children.
+ * Only emits events for keyed nodes.
+ *
+ * @param {Node} node - the initial node.
+ * @return {Node}
+ */
+function dispatch (node, type) {
+  // Trigger event for this element if it has a key.
+  if (getKey(node)) {
+    var ev = document.createEvent('Event')
+    var prop = { value: node }
+    ev.initEvent(type, false, false)
+    Object.defineProperty(ev, 'target', prop)
+    Object.defineProperty(ev, 'srcElement', prop)
+    node.dispatchEvent(ev)
+  }
+
+  // Dispatch to all children.
+  var child = node.firstChild
+  while (child) child = dispatch(child, type).nextSibling
+  return node
 }
 
 /**
@@ -280,8 +312,8 @@ function dispatch (el, type) {
  * Confirm that a value is truthy, throws an error message otherwise.
  *
  * @param {*} val - the val to test.
- * @param {String} msg - the error message on failure.
- * @throws Error
+ * @param {string} msg - the error message on failure.
+ * @throws {Error}
  */
 function assert (val, msg) {
   if (!val) throw new Error('set-dom: ' + msg)